refactor(parser): extract import-alias tracking into types/imports
Step 1 of the `parser/types.rs` split called out in audit Plan 2
(1376 lines, 38 top-level items). Move `TypeImports`,
`SourcesImports`, and the small `peel_annotated_inner` helper they
share into `crates/toolr-core/src/parser/types/imports.rs`.

`git mv types.rs types/mod.rs` — types/ is now a directory module.
New `types/imports.rs` (256 lines) holds the two structs cleanly
self-contained (only `ruff_python_ast::{Expr, ModModule, Stmt}`).
`types/mod.rs` shrinks 1376 → 1137 lines. Two private methods on
`TypeImports` (`resolve_direct`, `resolve_attribute`) get promoted
to `pub(super)` so the resolver in mod.rs can still reach them.

Public surface unchanged: parser::types::TypeImports and
parser::types::SourcesImports re-exported via pub use imports::*;.

Tests: 24 parser::types::tests::* pass; cargo check --workspace
clean.

Note: `--no-verify` used on this commit only — prek's commit-time
working-tree snapshot leaves both `types.rs` and `types/mod.rs` on
disk during its clippy hook (causing E0761), even though the staged
state correctly deletes the former. `prek run --all-files` ran
clean against the post-commit tree.

Remaining splits for follow-up commits / PRs:
- types/path_constraints.rs (PathConstraints + helpers)
- types/supported.rs (SupportedType + UnsupportedType + TypeResolutionError)
- types/literals.rs (literal_str / literal_str_list / literal_u32)
- types/arg_metadata.rs (extract_arg_metadata + section helpers)
- types/resolve.rs (resolve_arguments / resolve / friends)
